Colocando a DTO, arrumando o service e lançando algumas exceções

+++ sistemagendamento/src/main/java/com/sistemadecadastramento/services/UsuarioService.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.sistemadecadastramento.infra.UsuarioJaCadastradoException;
+import com.sistemadecadastramento.infra.UsuarioNaoExisteException;
 import com.sistemadecadastramento.models.Usuario;
 import com.sistemadecadastramento.repository.UsuarioRepository;
 
@@ -25,17 +27,29 @@ public class UsuarioService {
         return repository.findById(id);
     }
 
-    public Usuario salvar(Usuario usuario){
+    public Usuario salvarCriar(Usuario usuario){
+        if(repository.existsById(usuario.getId())){
+            throw new UsuarioJaCadastradoException();
+        }
         return repository.save(usuario);
     }
 
-    public boolean verificarUsuarioPorId(Long id){
-        // necessario implementar
+    public Usuario salvarAtualizar(Usuario usuario){
+        if(!existePorId(usuario.getId())){
+            throw new UsuarioNaoExisteException();
+        }
+
+        return repository.save(usuario);
+    }
+
+    public boolean existePorId(Long id){
         return repository.existsById(id);
     }
 
     public void deletar(Long id){
-        // necessario implementar
+        if(!repository.existsById(id)){
+            throw new UsuarioNaoExisteException();
+        }
         repository.deleteById(id);
     }
 }
